Day 1 Part 1&2 Solution
This was a bunch of iteration on part 2. Part 1 was a bit easier to
reason about, normalize position, track each time it is 0.

Had to peek at some other solutions when I swore I was doing it right...
but I was doing it right? Using the euclidian mod was a nice find and
once I refactored to do both in the same function it started working
again.

Oh well...
Solution:
```
=== AoC Day 1 Part 1 ===
Final Zero Counts: 1023

=== AoC Day 1 Part 2 ===
Final Zero Counts: 5899
```